Converted brain from continuous to atomic state.
Whereas previously, the program looped through the dictionary of
neurons and would accumulate weights continuously, but cause certain
neurons to exceed the threshold without firing, and firing would occur
based on the order of the neurons in the dictionary, this edit splits
the functioning of the brain into two time states.

The states are thisState, and nextState. This state always contains the
neurons that are firing now. nextState is then accumulated. At the end
of the reading of thisState (all neurons exceeding the threshold have
fired into nextState), thisState is switched to nextState, and any
neurons now exceeding the threshold as a result of the previous
thisState firing into the current thisState are now fired.