Final Logic Errors in Muscle firing correct
Previously, when the weight of a muscle neuron was accumulated, both
the values for this state and next state were set to 0. This would
prevent the muscle weighs from properly accumulating due to the
algorithm for the next state being equal to the accumulation plus the
current weight in this state.

Now, only this state is set to 0 at the end of the accumulation.