Inclusion of experimental optimization algorithm
This commit includes a highly experimental algorithm I hypothesized
would optimize the firing of neurons in an extremely scalable way.

A function, incrementNeuron(neuronName, weight), takes a neuron name
and weight. It runs the determined increment, and then checks if that
neuron’s next state exceeds the threshold.

If it does, then that neuron’s name is added to an array of neurons to
fire in the next iteration (the next state). When the next iteration
begins, only the neurons from the array are evaluated (fired) from the
dictionary, whereas before the entire dictionary was read each loop.

Theoretically, this should be helpful in increasingly scalable
scenarios, where thousands, millions, or billions of neurons are
involved, but only a small fraction of them are going to ever fire in a
particular time state.