Fixed experimental optimization
Experimental Optimization previously made same mistake as v7 algorithm,
which would fire its neurons into the same array before the array was
finished being iterated through, causing additional neurons to fire in
a given time state that were not meant to.

This is now solved by adding a boolean column to the dictionary that
says if the neuron that exceeds the threshold has or has not already
been added to the array of neurons to fire in the next state.

This update also removes the check for if the neuron exceeds the
threshold in the runconnectome() loop. This is because this check is
now done in the incrementNeuron() function where the neuron is added to
the list of neurons to fire in the next state, and thus does not have
to be checked twice. (special thanks to Geoffrey Churchill)