Neurons no longer overwrite, muscles explicitly listed.
In previous edition, the firing of two neurons into the same neuron
would cause the second neuron to overwrite the contribution of the
first neuron, a basic logical error.

Also, during the accumulation in the muscle neurons, each neuron would
be explicitly checked, whereas now we explicitly only check the left
and right muscle neurons.