Removed unnecessary files
Removed files that were overly experimental or were for presentations.
EO is for experimental optimization where neurons destined to fire in
the next time state are explicitly added to a list. TSEO files are
renamed to disembodiedConnectome.py and connectome.py. These should be
the standard running files because they are the most logically correct
ones.